+++ packages/http-client-java/emitter/src/code-model-builder.ts
@@ -966,29 +966,49 @@ export class CodeModelBuilder {
     responses: SdkHttpResponse[],
     sdkMethod: SdkMethod<SdkHttpOperation>,
   ) {
-    if (!this.isBranded()) {
-      // TODO: currently unbranded does not support paged operation
-      return;
-    }
-
     if (sdkMethod.kind === "paging" || sdkMethod.kind === "lropaging") {
       for (const response of responses) {
         const bodyType = response.type;
         if (bodyType && bodyType.kind === "model") {
-          const itemName = sdkMethod.response.resultPath;
-          const nextLinkName = sdkMethod.nextLinkPath;
+          const itemClientName = sdkMethod.response.resultPath;
+          const nextLinkClientName = sdkMethod.nextLinkPath;
 
-          op.extensions = op.extensions ?? {};
-          op.extensions["x-ms-pageable"] = {
-            itemName: itemName,
-            nextLinkName: nextLinkName,
-          };
+          let itemSerializedName: string | undefined = undefined;
+          let nextLinkSerializedName: string | undefined = undefined;
 
           op.responses?.forEach((r) => {
             if (r instanceof SchemaResponse) {
               this.trackSchemaUsage(r.schema, { usage: [SchemaContext.Paged] });
+
+              // find serializedName for items and nextLink
+              if (r.schema instanceof ObjectSchema) {
+                r.schema.properties?.forEach((p) => {
+                  if (
+                    itemClientName &&
+                    !itemSerializedName &&
+                    p.serializedName &&
+                    p.language.default.name === itemClientName
+                  ) {
+                    itemSerializedName = p.serializedName;
+                  }
+                  if (
+                    nextLinkClientName &&
+                    !nextLinkSerializedName &&
+                    p.serializedName &&
+                    p.language.default.name === nextLinkClientName
+                  ) {
+                    nextLinkSerializedName = p.serializedName;
+                  }
+                });
+              }
             }
           });
+
+          op.extensions = op.extensions ?? {};
+          op.extensions["x-ms-pageable"] = {
+            itemName: itemSerializedName,
+            nextLinkName: nextLinkSerializedName,
+          };
           break;
         }
       }
@@ -1054,8 +1074,21 @@ export class CodeModelBuilder {
           lroMetadata.finalStep.kind === "pollingSuccessProperty" &&
           lroMetadata.finalResponse.resultPath
         ) {
-          // final result is the value in lroMetadata.finalStep.target
-          finalResultPropertySerializedName = lroMetadata.finalResponse.resultPath;
+          const finalResultPropertyClientName = lroMetadata.finalResponse.resultPath;
+
+          // find serializedName for lro result
+          if (finalResultPropertyClientName) {
+            lroMetadata.finalResponse.envelopeResult.properties?.forEach((p) => {
+              // TODO: "p.__raw?.name" should be "p.name", after TCGC fix https://github.com/Azure/typespec-azure/issues/2072
+              if (
+                p.kind === "property" &&
+                p.serializedName &&
+                p.__raw?.name === finalResultPropertyClientName
+              ) {
+                finalResultPropertySerializedName = p.serializedName;
+              }
+            });
+          }
         }
       }