+++ packages/http-client-python/generator/pygen/codegen/models/operation_group.py
@@ -159,19 +159,23 @@ class OperationGroup(BaseModel):
                 ImportType.SDKCORE,
             )
             file_import.add_submodule_import(
-                ".._configuration",
+                self.code_model.get_relative_import_path(
+                    serialize_namespace,
+                    self.code_model.get_imported_namespace_for_client(self.client.client_namespace, async_mode),
+                    module_name="_configuration",
+                ),
                 f"{self.client.name}Configuration",
                 ImportType.LOCAL,
             )
             file_import.add_msrest_import(
-              serialize_namespace=kwargs.get("serialize_namespace", self.code_model.namespace),
-              msrest_import_type=MsrestImportType.Serializer,
-              typing_section=TypingSection.REGULAR,
+                serialize_namespace=kwargs.get("serialize_namespace", self.code_model.namespace),
+                msrest_import_type=MsrestImportType.Serializer,
+                typing_section=TypingSection.REGULAR,
             )
             file_import.add_msrest_import(
-              serialize_namespace=kwargs.get("serialize_namespace", self.code_model.namespace),
-              msrest_import_type=MsrestImportType.SerializerDeserializer,
-              typing_section=TypingSection.REGULAR,
+                serialize_namespace=kwargs.get("serialize_namespace", self.code_model.namespace),
+                msrest_import_type=MsrestImportType.SerializerDeserializer,
+                typing_section=TypingSection.REGULAR,
             )
         if self.has_abstract_operations:
             file_import.add_submodule_import(